Add files via upload
This text adventure game is based on the course Data Science in Python with Professor Kusterer. The target audience is thereby Business Analytics students, and others with an interest in Business Analytics. The game is inspired by random quotas from Chase Kusterer while preparing his students for a year with coding.  

The code contains:
Strings and String Manipulation
User-Defined Functions and Objects 
User Input and Input Handling
Conditional and Nested Conditional Statements 
Loops 
Controlling for Errors